adjusting the structure design of the webpage
