+++ src/Login.html
@@ -1,0 +1,234 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>Socialise</title>
+    <link href="output.css" rel="stylesheet">
+    <style>
+        .button-black-text {
+            background-color: #e5e7eb;
+            color: black;
+            padding: 0.5rem;
+            border-radius: 0.375rem;
+        }
+
+        .navbar-input {
+            width: 100%;
+            max-width: 600px;
+        }
+
+        .input {
+            background-color: #f1f5f9;
+            color: #fd0707;
+            width: 50%;
+            max-width: 500px;
+        }
+
+        .post-card {
+            background-color: #f9fafb;
+            border-radius: 0.375rem;
+            padding: 1rem;
+            margin-bottom: 1rem;
+            box-shadow: 0 1px 2px rgba(0, 0, 0, 0.05);
+            display: flex;
+            flex-direction: column;
+        }
+
+        .post-header {
+            display: flex;
+            align-items: center;
+            margin-bottom: 0.5rem;
+        }
+
+        .post-header img {
+            border-radius: 50%;
+            margin-right: 0.5rem;
+            width: 32px;
+            height: 32px;
+        }
+
+        .post-meta {
+            color: #9ca3af;
+            font-size: 0.875rem;
+        }
+
+        .post-title {
+            font-weight: bold;
+            margin-bottom: 0.5rem;
+            font-size: 1.125rem;
+        }
+
+        .post-body {
+            color: #4b5563;
+            font-size: 0.875rem;
+        }
+
+        .badge-container {
+            display: flex;
+            gap: 0.5rem;
+            margin-bottom: 0.5rem;
+        }
+
+        .badge {
+            background-color: #e5e7eb;
+            color: #1f2937;
+            padding: 0.25rem 0.5rem;
+            border-radius: 9999px;
+            font-size: 0.75rem;
+            text-decoration: none;
+        }
+
+        .badge:hover {
+            background-color: #d1d5db;
+        }
+
+        .sidebar {
+            flex: 1;
+            padding: 1rem;
+        }
+
+        .create-discussion-card {
+            background-color: #fff;
+            border-radius: 0.375rem;
+            padding: 1rem;
+            box-shadow: 0 1px 2px rgba(0, 0, 0, 0.05);
+            margin-bottom: 1rem;
+            display: flex;
+            flex-direction: column;
+            align-items: center;
+        }
+
+        .create-discussion-button {
+            background-color: #6366f1;
+            color: white;
+            padding: 0.5rem 1rem;
+            border-radius: 0.375rem;
+            margin-top: 1rem;
+            text-align: center;
+        }
+
+        .right-banner {
+            background-color: white;
+            border-radius: 0.375rem;
+            padding: 1rem;
+            box-shadow: 0 1px 2px rgba(0, 0, 0, 0.05);
+            margin-left: 1rem;
+            width: 250px; /* Adjust the width to fit better */
+        }
+
+        .right-banner img {
+            border-radius: 0.375rem;
+            width: 100%;
+            height: auto;
+        }
+
+        .right-banner p {
+            margin: 1rem 0;
+            font-size: 0.875rem;
+            color: #4b5563;
+        }
+
+        .right-banner .create-discussion-button {
+            margin-top: 0;
+        }
+
+        .overlay {
+            position: fixed;
+            top: 0;
+            left: 0;
+            width: 100%;
+            height: 100%;
+            background-color: transparent;
+            display: flex;
+            justify-content: center;
+            align-items: center;
+            z-index: 1000;
+        }
+
+        .login-popup {
+            background-color: white;
+            padding: 2rem;
+            border-radius: 0.375rem;
+            box-shadow: 0 10px 15px rgba(0, 0, 0, 0.1);
+            width: 400px;
+            max-width: 90%;
+        }
+
+        .login-popup input {
+            width: 100%;
+            padding: 0.75rem;
+            margin-bottom: 1rem;
+            border: 1px solid #d1d5db;
+            border-radius: 0.375rem;
+        }
+
+        .login-popup button {
+            background-color: #6366f1;
+            color: white;
+            padding: 0.75rem;
+            width: 100%;
+            border: none;
+            border-radius: 0.375rem;
+            cursor: pointer;
+        }
+
+        .login-popup a {
+            color: #6366f1;
+            text-decoration: none;
+            display: block;
+            text-align: center;
+            margin-top: 0.5rem;
+        }
+
+        .login-popup a:hover {
+            text-decoration: underline;
+        }
+    </style>
+</head>
+
+<body class="h-max bg-gray-100">
+    <div class="flex justify-between items-center p-4">
+        <img src="../images/socialise2.png" alt="socialise" class="h-8 ml-4">
+        <div class="flex justify-end w-1/4">
+            <button class="btn btn-sm mr-4">Login</button>
+            <button class="btn btn-sm mr-4">Sign Up</button>
+        </div>
+    </div>
+
+    <!-- Overlay and Login Popup -->
+    <div class="overlay">
+        <div class="login-popup">
+            <h2 class="text-2xl font-bold mb-4 text-center">Log In</h2>
+            <input type="text" placeholder="Email or Username" autocomplete="off" id="username" name="username" />
+            <input type="password" placeholder="Password" autocomplete="new-password" id="password" name="password" />
+            <button id="loginButton">Log In</button>
+            <div id="loginMessage" class="text-center mt-4"></div> <!-- Added message div -->
+            <a href="#">Forgot password?</a>
+            <a href="#">New to Socialise? Sign Up</a>
+        </div>
+    </div>
+
+    <script>
+        document.getElementById('loginButton').addEventListener('click', async function() {
+        const usernameOrEmail = document.getElementById('username').value;
+        const password = document.getElementById('password').value;
+        const loginMessage = document.getElementById('loginMessage');
+
+        const response = await fetch(`http://localhost:3000/login?usernameOrEmail=${usernameOrEmail}&password=${password}`, {
+            method: 'GET'
+        });
+
+        if (response.ok) {
+            const userData = await response.json();
+            loginMessage.innerHTML = '<p class="text-green-600">Login successful!</p>';
+            // Redirect to profile page with username and email
+            window.location.href = `profile.html?username=${userData.username}&email=${userData.email}`;
+        } else {
+            const errorMessage = await response.text();
+            loginMessage.innerHTML = `<p class="text-red-600">${errorMessage}</p>`;
+        }
+    });
+    </script>
+</body>
+</html>
