+++ src/profile.html
@@ -153,69 +153,37 @@
         </div>
     </header>
 
-    <div class="flex flex-1">
-        <!-- Sidebar -->
-        <aside class="w-72 bg-white p-4 border-r-2 flex flex-col">
-            <nav class="flex-1">
-                <ul>
-                    <li class="mb-4">
-                        <a href="#" class="flex items-center text-gray-800">
-                            <span class="mr-2">🏠</span> Home
-                        </a>
-                    </li>
-                    <li class="mb-4">
-                        <a href="#" class="flex items-center text-gray-800">
-                            <span class="mr-2">👤</span> Profile
-                        </a>
-                    </li>
-                    <li class="mb-8">
-                        <a href="#" class="flex items-center text-gray-800">
-                            <span class="mr-2">📰</span> News
-                        </a>
-                    </li>
-                </ul>
-                <div class="mb-8">
-                    <h2 class="text-gray-600 mb-2">FAVOURITES</h2>
-                    <ul>
-                        <li class="mb-2">
-                            <a href="#" class="flex items-center text-gray-800">
-                                <span class="mr-2">🩺</span> Healthcare
-                            </a>
-                        </li>
-                        <li class="mb-2">
-                            <a href="#" class="flex items-center text-gray-800">
-                                <span class="mr-2">🌍</span> Climate
-                            </a>
-                        </li>
-                        <li>
-                            <a href="#" class="flex items-center text-gray-800">
-                                <span class="mr-2">📚</span> Education
-                            </a>
-                        </li>
-                    </ul>
+    <!-- main -->
+    <div class="flex flex-row h-full">
+        <!-- sidebar -->
+        <div class="flex flex-col items-center w-1/6 h-full bg-white border-r-2">
+            <div class="flex flex-col w-4/5 h-fit border-b-2 px-4 py-8 gap-2">
+                <div class="flex items-center   ">
+                    <button class="btn w-full bg-white border-0 text-left shadow-none"><span class="flex items-center w-full gap-2"><img src="../images/home-variant-outline.svg" width="30px" />Home</span></button>
                 </div>
-                <div>
-                    <h2 class="text-gray-600 mb-2">LEARN MORE</h2>
-                    <ul>
-                        <li class="mb-2">
-                            <a href="#" class="flex items-center text-gray-800">
-                                <span class="mr-2">ℹ️</span> About Us
-                            </a>
-                        </li>
-                        <li class="mb-2">
-                            <a href="#" class="flex items-center text-gray-800">
-                                <span class="mr-2">✉️</span> Contact Us
-                            </a>
-                        </li>
-                        <li>
-                            <a href="#" class="flex items-center text-gray-800">
-                                <span class="mr-2">📊</span> User Statistics
-                            </a>
-                        </li>
-                    </ul>
+                <div class="flex items-center   ">
+                    <button class="btn w-full bg-white border-0 text-left shadow-none"><span class="flex items-center w-full gap-2"><img src="../images/account-circle-outline.svg" width="30px" />Profile</span></button>
                 </div>
-            </nav>
-        </aside>
+                <div class="flex items-center   ">
+                    <button class="btn w-full bg-white border-0 text-left shadow-none"><span class="flex items-center w-full gap-2"><img src="../images/newspaper-variant-outline.svg" width="30px" />News</span></button>
+                </div>
+            </div>
+            <div class="flex flex-col w-4/5 h-fit border-b-2 px-4 py-8 gap-2">
+                <h2 class="font-bold mb-2">FOR YOU</h2>
+            </div>
+            <div class="flex flex-col w-4/5 h-fit border-b-2 px-4 py-8 gap-2">
+                <h2 class="font-bold mb-2">LEARN MORE</h2>
+                <div class="flex items-center   ">
+                    <button class="btn w-full bg-white border-0 text-left shadow-none"><span class="flex items-center w-full gap-2"><img src="../images/information-outline.svg" width="30px" />About Us</span></button>
+                </div>
+                <div class="flex items-center   ">
+                    <button class="btn w-full bg-white border-0 text-left shadow-none"><span class="flex items-center w-full gap-2"><img src="../images/phone-in-talk.svg" width="30px" />Contact Us</span></button>
+                </div>
+                <div class="flex items-center   ">
+                    <button class="btn w-full bg-white border-0 text-left shadow-none"><span class="flex items-center w-full gap-2"><img src="../images/finance.svg" width="30px" />User Statistics</span></button>
+                </div>
+            </div>
+        </div>
 
         <!-- Main Content -->
         <main class="flex-1 p-8">
