+++ src/signup.html
@@ -31,9 +31,13 @@
         }
     </style>
 </head>
-<body class="bg-gray-100">
+<body class="h-max bg-gray-100">
     <div class="flex justify-between items-center p-4">
         <img src="../images/socialise2.png" alt="socialise" class="h-8 ml-4">
+        <div class="flex justify-end w-1/4">
+            <button class="btn btn-sm mr-4">Login</button>
+            <button class="btn btn-sm mr-4">Sign Up</button>
+        </div>
     </div>
     <div class="flex items-center justify-center content">
         <div class="w-full max-w-md bg-white shadow-md rounded-2xl p-8">
@@ -42,6 +46,7 @@
                 <h1 class="text-2xl font-bold text-gray-800">Sign Up</h1>
                 <p class="text-gray-600">Join Us in the Discussion!</p>
             </div>
+            
             <form id="signupForm" autocomplete="off">
                 <!-- Hidden dummy input -->
                 <input type="text" style="display:none" autocomplete="false">
